+++ src/timeatlas/__init__.py
@@ -1,16 +1,17 @@
 """ TimeAtlas API Definition """
 
 from timeatlas import abstract
+from timeatlas import detectors
+from timeatlas import generators
 from timeatlas.metadata import Metadata
-from timeatlas.time_series import TimeSeries
-from timeatlas.time_series_dataset import TimeSeriesDataset
 from timeatlas import metrics
+from timeatlas import models
 from timeatlas import plots
 from timeatlas import processors
-from timeatlas import types
 from timeatlas.read_write import (
     read_pickle,
     read_text
 )
-
-from timeatlas.generators import AnomalyGenerator, AnomalyGeneratorTemplate
+from timeatlas.time_series import TimeSeries
+from timeatlas.time_series_dataset import TimeSeriesDataset
+from timeatlas import types
