+++ src/timeatlas/detectors/__init__.py
@@ -1,5 +1,5 @@
 from .surprise import Surprise
 
 __all__ = [
-    "Surprise",
+    'Surprise',
 ]
