+++ src/timeatlas/metadata/__init__.py
@@ -1,1 +1,5 @@
 from .metadata import Metadata
+
+__all__ = [
+    'Metadata'
+]
