+++ src/timeatlas/read_write/__init__.py
@@ -1,8 +1,15 @@
-from .text import read_text, csv_to_tsd
-from .pickle import read_pickle
+from .pickle import (
+    read_pickle
+)
+
+from .text import (
+    read_text,
+    csv_to_tsd
+)
+
 
 __all__ = [
-    "read_text",
     "read_pickle",
-    "csv_to_tsd"
+    "csv_to_tsd",
+    "read_text"
 ]
