+++ src/timeatlas/utils/utils.py
@@ -1,7 +1,6 @@
 import os
 from typing import NoReturn, Any
 import pickle
-from matplotlib.axes import Axes
 
 
 def ensure_dir(file_path: str) -> NoReturn:
@@ -27,14 +26,3 @@ def to_pickle(obj: Any, path: str) -> NoReturn:
     ensure_dir(path)
     with open(path, 'wb') as file:
         pickle.dump(obj, file, pickle.HIGHEST_PROTOCOL)
-
-
-def add_metadata_to_plot(meta, ax: Axes):
-    if meta is not None:
-        if "unit" in meta:
-            unit = meta["unit"]
-            ax.set_ylabel("{} $[{}]$".format(unit.name, unit.symbol))
-        if "sensor" in meta:
-            sensor = meta["sensor"]
-            ax.set_title("{}—{}".format(sensor.id, sensor.name))
-    return ax
